:fix: missing include

+++ lib/headers/my.h
@@ -10,6 +10,7 @@
     #define _LOW_CHAR_ (str[index - 1] >= 'a' && str[index - 1] <= 'z')
     #define _NUM_ (str[index - 1] >= '0' && str[index - 1] <= '9')
     #include <stdlib.h>
+
 void my_putchar(char c);
 char *my_strstr(char *str, char const *to_find);
 int my_isneg(int nb);
@@ -52,5 +53,5 @@ char *my_inttos(int nbr);
 int my_intlen(int nbr);
 int my_printf(const char *format, ...);
 void my_puterror(char *str);
-void do_strmalloc(char *str, ssize_t size);
+void do_strmalloc(char *str, long size);
 #endif
